final project with changes and tweaks to the functionality and UI of the app
