+++ package.json
@@ -7,12 +7,14 @@
     "ajv-errors": "^1.0.1",
     "axios": "^0.27.2",
     "babel-preset-es2015": "^6.24.1",
+    "prisma": "^4.4.0",
     "prop-types": "^15.8.1",
     "react": "^18.1.0",
     "react-dom": "^18.1.0",
     "react-icons": "^4.4.0",
-    "react-router-dom": "^6.3.0",
+    "react-router-dom": "^5.3.4",
     "react-scripts": "^2.1.3",
+    "styled-components": "^5.3.6",
     "url-loader": "^4.1.1"
   },
   "scripts": {
