--- src/TodoList.js
@@ -1,19 +1,0 @@
-import React from "react";
-import TodoListItem from "./TodoListItem";
-
-function TodoList({ todoList, onRemoveTodo }) {
-  return (
-    <ul>
-      {todoList.map((todo) => (
-        <TodoListItem
-          key={todo.id}
-          todo={todo}
-          // title={todo.title}
-          onRemoveTodo={onRemoveTodo}
-        />
-      ))}
-    </ul>
-  );
-}
-
-export default TodoList;
