+++ src/components/Footer.js
@@ -1,13 +1,17 @@
 import React from "react";
 import style from "../css/Footer.module.css";
-import About from "./About";
+import { Link } from "react-router-dom";
 
 const Footer = () => {
+  const today = new Date();
+  const thisYear = today.getFullYear();
   return (
     <div>
       <footer className={style.footer}>
-        <About />
-        <p>Copyright &copy; 2022</p>
+        <p> &copy; Erick Odero {thisYear}</p>
+        <Link to="/about" className={style.link}>
+          About
+        </Link>
       </footer>
     </div>
   );
