+++ src/components/Landing.js
@@ -1,0 +1,12 @@
+import React from "react";
+import style from "../css/landing.module.css";
+
+function Landing() {
+  return (
+    <div>
+      <button className={style.landing_page}>Start</button>
+    </div>
+  );
+}
+
+export default Landing;
