+++ src/css/AddTodoForm.module.css
@@ -4,10 +4,19 @@
 }
 
 form {
-  display: flex;
+  display: inline-block;
   flex-direction: column;
-  width: 376px;
-  margin-left: 4px;
+  width: 100%;
+  border-bottom: 2px solid #fafafa;
+  margin-left: 80px;
+  /* margin-left: 4px; */
+  margin-top: 0;
+}
+
+input {
+  width: 200px;
+  height: 25px;
+  padding-right: 20px;
 }
 .title {
   color: black;
@@ -16,20 +25,21 @@ form {
 .button {
   text-align: center;
   background: green;
-  width: 100%;
+  width: 60px;
   height: 30px;
   color: #fff;
   font-weight: bold;
 
   border: none;
-  margin: 8px 0 5px 0;
+  margin-left: 68px;
   border-radius: 3px;
   font-size: 0.8rem;
   cursor: pointer;
+  align-self: center;
 }
 
 @media screen and (max-width: 445px) {
   form {
-    width: 95%;
+    width: 60%;
   }
 }
