+++ src/css/App.module.css
@@ -1,28 +1,54 @@
-@import url("https://fonts.googleapis.com/css2?family=Poppins:wght@300;400&display=swap");
+@import url("https://fonts.googleapis.com/css2?family=Quicksand:wght@300;400;500;600;700&display=swap");
+
+* {
+  margin: 0;
+  font-family: "Quicksand";
+  color: #333;
+}
+
+.content {
+  max-width: 600px;
+  margin: 1px auto 0px;
+  padding: 20px;
+  border-left: 1px solid black;
+  border-right: 1px solid black;
+}
+/* @media screen and (max-width: 600px) {
+  .content {
+    height: 100%;
+  } */
+/* } */
+@media screen and (max-width: 640px) {
+  .content {
+    width: fit-content;
+  }
+}
+
+/* @import url("https://fonts.googleapis.com/css2?family=Poppins:wght@300;400&display=swap");
 
 * {
   box-sizing: border-box;
   font-family: poppins, sans-serif;
   margin: 0;
   padding: 0;
-}
-
-.Container {
-  display: flex;
-  /* align-items: center; */
-  flex-direction: column;
-  min-height: 700px;
-  max-width: 450px;
-  /* margin: 30px auto; */
-  margin: 1px auto;
+} */
+/*  */
+/* .Container { */
+/* display: flex; */
+/* align-items: center; */
+/* flex-direction: column; */
+/* min-height: 700px; */
+/* max-width: 450px; */
+/* margin: 30px auto; */
+/* margin: 1px auto;
   border: 2px solid black;
   box-shadow: 0 2.5px rgba(104, 104, 104, 0.8);
   padding: 30px;
   border-radius: 5px;
-  background-color: rgb(228, 216, 202);
+  background-color: rgb(235, 217, 196);
   color: #fffeef;
-}
-.about_image {
+} */
+/* .about_image {
   display: inline-flex;
   color: black;
   text-decoration: none;
@@ -30,14 +56,13 @@
   align-self: center;
   height: 100px;
   width: 100px;
-}
+} */
 
-.loading {
+/* .loading {
   color: black;
 }
 
 @media screen and (max-width: 600px) {
   .container {
     height: 100%;
-  }
-}
+  /* } */
