+++ src/css/Footer.module.css
@@ -1,15 +1,44 @@
 .footer {
   display: flex;
   flex-direction: column;
+  justify-content: center;
   align-items: center;
   color: black;
   position: relative;
   bottom: 0;
-  margin-top: auto;
+  background-color: #f2f2f2;
+
+  max-width: 640px;
+  height: 100px;
+  /* margin-top: auto; */
+  margin: auto;
+  border-left: 1px solid black;
+  border-right: 1px solid black;
+  border-left: 1px solid black;
+  border-bottom: 1px solid black;
 }
 
-.paragraph {
-  font-size: 0.8rem;
+.about {
+  display: flex;
+  color: black;
+  margin: 30px auto;
+  font-size: 1rem;
+  width: 350px;
+  border: 2px solid #f2f2f2;
+}
+
+.about a:hover {
+  color: #e9d18e;
+}
+.link,
+.back {
+  text-decoration: none;
+  color: black;
+}
+
+.back {
+  margin-left: 850px;
+  color: rebeccapurple;
 }
 
 @media screen and (max-width: 600px) {
