+++ src/css/Navbar.module.css
@@ -1,0 +1,53 @@
+.navbar {
+  padding: 20px;
+  display: flex;
+  background-color: #f2f2f2;
+  align-items: center;
+  width: 600px;
+  margin: 0 auto;
+  border-bottom: 0.5px solid #f2f2f2;
+  border-left: 1px solid black;
+  border-left: 1px solid black;
+  border-right: 1px solid black;
+}
+
+.navbar h1 {
+  color: black;
+}
+
+.navbar .links {
+  margin-left: auto;
+  color: black;
+  font-weight: bold;
+}
+
+.navbar a {
+  margin-left: 16px;
+  text-decoration: none;
+  padding: 6px;
+}
+
+.navbar a:hover {
+  color: #e9d18e;
+}
+
+@media screen and (max-width: 640px) {
+  .navbar {
+    padding: 2px;
+    display: flex;
+    flex-direction: column;
+    background-color: #f2f2f2;
+    justify-items: center;
+    align-items: center;
+    width: fit-content;
+  }
+  .navbar h1 {
+    color: black;
+    font-size: medium;
+  }
+
+  .navbar .links {
+    color: black;
+    font-weight: bold;
+  }
+}
